Implement merged queries.
Summary: Implement merged queries.

Test Plan:
Run a query in one place, and run an exact clone of it at the same time. Ensure that both query tables are equal.
Stop one query, ensure the first one still is updating. Stop the other query, and ensure that the state is cleared and the query is no longer executing.

Reviewers: samrg472

Reviewed By: samrg472

Differential Revision: https://p.dev.dglogik.com/D6